Use closed data in WORCS project

This function saves a data.frame as a .csv file (using write.csv), stores a checksum in '.worcs', appends the .gitignore file to exclude filename, and saves a synthetic copy of data for public use. To generate these synthetic data, the function synthetic is used.

closed_data(
  data,
  filename = paste0(deparse(substitute(data)), ".csv"),
  codebook = paste0("codebook_", deparse(substitute(data)), ".Rmd"),
  worcs_directory = ".",
  synthetic = TRUE,
  ...
)

Arguments

data
A data.frame to save.

filename
Character, naming the file data should be written to. By default, constructs a filename from the name of the object passed to data.

codebook
Character, naming the file the codebook should be written to. An 'R Markdown' codebook will be created and rendered to github_document ('markdown' for 'GitHub'). By default, constructs a filename from the name of the object passed to data, adding the word 'codebook'. Set this argument to NULL to avoid creating a codebook.

worcs_directory
Character, indicating the WORCS project directory to which to save data. The default value "." points to the current directory.

synthetic
Logical, indicating whether or not to create a synthetic dataset using the synthetic function. Additional arguments for the call to synthetic can be passed through ....

...
Additional arguments passed to and from functions.

Value

Returns NULL invisibly. This function is called for its side effects.

See also

Examples

old_wd <- getwd()
test_dir <- file.path(tempdir(), "data")
dir.create(test_dir)
setwd(test_dir)
worcs:::write_worcsfile(".worcs")
df <- iris[1:3, ]
closed_data(df, codebook = NULL)

#> v  Storing original data in 'df.csv' and updating the checksum in '.worcs'.
#> v  Creating C:/Users/lissa102/AppData/Local/Temp/RtmpA5EzTd/data/README.md.
#> v  Generating synthetic data for public use. Ensure that no identifying information is included.
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%

#> Warning: Dropped unused factor level(s) in dependent variable: versicolor, virginica.

#>   |                                                                              |======================================================================| 100%
#> v  Storing synthetic data in 'synthetic_df.csv' and updating the checksum in '.worcs'.
#> v  Updating '.gitignore'.
#> v  Updating C:/Users/lissa102/AppData/Local/Temp/RtmpA5EzTd/data/README.md.
#> v  Updating '.gitignore'.

setwd(old_wd)
unlink(test_dir, recursive = TRUE)