Load project entry points

Loads the designated project entry point into the default editor, using file.edit.

Usage

load_entrypoint(worcs_directory = ".", verbose = TRUE, ...)

Arguments

worcs_directory

Character, indicating the WORCS project directory to which to save data. The default value "." points to the current directory.

verbose

Logical. Whether or not to print status messages to the console. Default: TRUE

...

Additional arguments passed to file.edit.

Value

No return value. This function is called for its side effects.

Examples

if (FALSE) {
# Create directory to run the example
old_wd <- getwd()
test_dir <- file.path(tempdir(), "entrypoint")
dir.create(test_dir)
setwd(test_dir)
# Prepare worcs file and dummy entry point
worcs:::write_worcsfile(".worcs", entry_point = "test.txt")
writeLines("Hello world", con = file("test.txt", "w"))
# Demonstrate load_entrypoint()
load_entrypoint()
# Cleaning example directory
setwd(old_wd)
unlink(test_dir, recursive = TRUE)
}