Save open data file

This function saves a data.frame as a .csv file, stores a checksum in 'data_checksum.txt', appends the .gitignore file to exclude 'data.csv', and writes a stem for 'data_cleaning.R'.

open_data(data, codebook = TRUE)

Arguments

data
A data.frame to save.

codebook
Logical, indicating whether to render a codebook or not. If set to TRUE, the default, a file called 'codebook.Rmd' is created and rendered to 'codebook.md' for 'GitHub'.

Value

Returns NULL invisibly. This function is called for its side effects.

See also

Examples

the_test <- "opendata"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
setwd(file.path(tempdir(), the_test))
open_data(iris[1:5, ], codebook = NULL)

#> Could not find .gitignore file. You might not be working from the main project directory. Created .gitignore file in folder: C:/Users/lissa102/AppData/Local/Temp/RtmpIxfj49/opendata

#> 
#> 
#> processing file: codebook.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |..............                                                        |  20%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: unnamed-chunk-1
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: codebook.knit.md

#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS codebook.utf8.md --to gfm --from markdown+autolink_bare_uris+tex_math_single_backslash --output codebook.md --standalone --template "C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\default.md" 
#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS codebook.md --to html4 --from gfm --output codebook.html --standalone --self-contained --highlight-style pygments --template "C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\PREVIE~1.HTM" --variable "github-markdown-css:C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\github.css" --email-obfuscation none --metadata pagetitle=PREVIEW 

#> 
#> Preview created: codebook.html

#> 
#> Output created: codebook.md

#> Storing original data in 'data.csv' and updating the checksum in '.worcs'.

#> No '.worcs' file found; either this is not a worcs project, or the working directory is not set to the project directory.

#> Updating ".gitignore".

#> Generating "data_cleaning.R"

setwd(old_wd)
unlink(file.path(tempdir(), the_test))