Snapshot endpoints in WORCS project

Update the checksums of all endpoints in a WORCS project.

snapshot_endpoints(worcs_directory = ".", verbose = TRUE, ...)

Arguments

worcs_directory

Character, indicating the WORCS project directory to which to save data. The default value "." points to the current directory. Default: '.'

verbose

Logical. Whether or not to print status messages to the console. Default: TRUE

...

Additional arguments.

Value

No return value. This function is called for its side effects.

See also

Examples

# Create directory to run the example
old_wd <- getwd()
test_dir <- file.path(tempdir(), "update_endpoint")
dir.create(test_dir)
setwd(test_dir)
file.create(".worcs")
#> [1] TRUE
writeLines("test", "test.txt")
add_endpoint("test.txt")
#> This worcs project is not version controlled with Git.
#> ✔ Adding endpoint 'test.txt' to '.worcs'.
writeLines("second test", "test.txt")
snapshot_endpoints()
#> This worcs project is not version controlled with Git.
#> ✔ Update snapshot of endpoint 'test.txt'.
# Cleaning example directory
setwd(old_wd)
unlink(test_dir, recursive = TRUE)