Essential citations Knit function for 'RStudio'

This is a wrapper for render. First, this function parses the citations in the document, removing citations marked with double at sign, e.g.: @@reference2020. Then, it renders the file.

cite_essential(...)

Arguments

...

All arguments are passed to render.

Value

Returns NULL invisibly. This function is called for its side effect of rendering an 'R Markdown' file.

Examples

# NOTE: Do not use this function interactively, as in the example below.
# Only specify it as custom knit function in an R Markdown file, like so:
# knit: worcs::cite_all

if (rmarkdown::pandoc_available("2.0")){
  file_name <- tempfile("citeessential", fileext = ".Rmd")
  rmarkdown::draft(file_name,
                   template = "github_document",
                   package = "rmarkdown",
                   create_dir = FALSE,
                   edit = FALSE)
  write(c("", "Optional reference: @reference2020"),
        file = file_name, append = TRUE)
  cite_essential(file_name)
}
#> 
#> 
#> processing file: citeessential689443b6499a.Rmd
#>   |                                                                              |                                                                      |   0%  |                                                                              |..........                                                            |  14%
#>   ordinary text without R code
#> 
#>   |                                                                              |....................                                                  |  29%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |..............................                                        |  43%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................                              |  57%
#> label: cars
#>   |                                                                              |..................................................                    |  71%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................................................          |  86%
#> label: pressure (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> output file: citeessential689443b6499a.knit.md
#> "C:/Program Files/RStudio/bin/quarto/bin/pandoc" +RTS -K512m -RTS citeessential689443b6499a.knit.md --to gfm+tex_math_dollars-yaml_metadata_block --from markdown+autolink_bare_uris+tex_math_single_backslash --output citeessential689443b6499a.md --template "C:\PROGRA~1\R\R-41~1.2\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\default.md" 
#> "C:/Program Files/RStudio/bin/quarto/bin/pandoc" +RTS -K512m -RTS citeessential689443b6499a.md --to html4 --from gfm+tex_math_dollars --output citeessential689443b6499a.html --standalone --self-contained --highlight-style pygments --template "C:\PROGRA~1\R\R-41~1.2\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\PREVIE~1.HTM" --variable "github-markdown-css:C:\PROGRA~1\R\R-41~1.2\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\github.css" --metadata pagetitle=PREVIEW --mathjax 
#> 
#> Preview created: citeessential689443b6499a.html
#> 
#> Output created: citeessential689443b6499a.md