Create codebook for a dataset

Creates a codebook for a dataset in 'R Markdown' format, and renders it to 'markdown' for 'GitHub'. Users can customize the 'R Markdown' document and re-knit it, for example, to add a paragraph with details on the data collection procedures. The variable descriptives are stored in a .csv file, which can be edited in 'R' or a spreadsheet program. Columns can be appended, and we encourage users to complete at least the following two columns in this file:

category Describe the type of variable in this column. For example: "morality".

description Provide a plain-text description of the variable. For example, the full text of a questionnaire item: "People should be willing to do anything to help a member of their family".

Re-knitting the 'R Markdown' file (using render) will transfer these changes to the 'markdown' file for 'GitHub'.

make_codebook(
  data,
  filename = "codebook.Rmd",
  render_file = TRUE,
  csv_file = gsub("Rmd$", "csv", filename)
)

Arguments

data
A data.frame for which to create a codebook.

filename
Character. File name to write the codebook rmarkdown file to.

render_file
Logical. Whether or not to render the document.

csv_file
Character. File name to write the codebook rmarkdown file to. By default, uses the filename stem of the filename argument. Set to NULL to write the codebook only to the 'R Markdown' file, and not to .csv.

Value

Logical, indicating whether or not the operation was successful. This function is mostly called for its side effect of rendering an 'R Markdown' codebook.

Examples

library(rmarkdown)
library(knitr)
filename <- tempfile("codebook", fileext = ".Rmd")
make_codebook(iris, filename = filename, csv_file = NULL)

#> 
#> 
#> processing file: codebook4e64304a2d25.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |..............                                                        |  20%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: unnamed-chunk-1
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: codebook4e64304a2d25.knit.md

#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS codebook4e64304a2d25.utf8.md --to gfm --from markdown+autolink_bare_uris+tex_math_single_backslash --output codebook4e64304a2d25.md --standalone --template "C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\default.md" 
#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS codebook4e64304a2d25.md --to html4 --from gfm --output codebook4e64304a2d25.html --standalone --self-contained --highlight-style pygments --template "C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\PREVIE~1.HTM" --variable "github-markdown-css:C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\github.css" --email-obfuscation none --metadata pagetitle=PREVIEW 

#> 
#> Preview created: codebook4e64304a2d25.html

#> 
#> Output created: codebook4e64304a2d25.md

#> [1] TRUE

unlink(c(
  ".worcs",
  filename,
  gsub("\\.Rmd", "\\.md", filename),
  gsub("\\.Rmd", "\\.html", filename),
  gsub("\\.Rmd", "_files", filename)
), recursive = TRUE)