Load data file

If the original data are available, this function loads the original data. If only a synthetic dataset is available, this function loads the synthetic data.

load_data()

Value

A data.frame of class "worcs_data", containing the original data, if available, or a synthetic version of the data, if the original data are unavailable. The data.frame has an attribute, "type", indicating whether the data are synthetic or original.

Examples

the_test <- "loaddata"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
setwd(file.path(tempdir(), the_test))
open_data(iris[1:5, ], codebook = NULL)

#> Could not find .gitignore file. You might not be working from the main project directory. Created .gitignore file in folder: C:/Users/lissa102/AppData/Local/Temp/RtmpIxfj49/loaddata

#> 
#> 
#> processing file: codebook.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |..............                                                        |  20%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: unnamed-chunk-1
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: codebook.knit.md

#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS codebook.utf8.md --to gfm --from markdown+autolink_bare_uris+tex_math_single_backslash --output codebook.md --standalone --template "C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\default.md" 
#> "C:/Program Files/RStudio/bin/pandoc/pandoc" +RTS -K512m -RTS codebook.md --to html4 --from gfm --output codebook.html --standalone --self-contained --highlight-style pygments --template "C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\PREVIE~1.HTM" --variable "github-markdown-css:C:\PROGRA~1\R\R-40~1.0\library\RMARKD~1\RMARKD~1\TEMPLA~1\GITHUB~1\RESOUR~1\github.css" --email-obfuscation none --metadata pagetitle=PREVIEW 

#> 
#> Preview created: codebook.html

#> 
#> Output created: codebook.md

#> Storing original data in 'data.csv' and updating the checksum in '.worcs'.

#> No '.worcs' file found; either this is not a worcs project, or the working directory is not set to the project directory.

#> Updating ".gitignore".

#> Generating "data_cleaning.R"

df <- load_data()
df

#> This is the original data set. The first 6 rows are:
#> 
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa

setwd(old_wd)
unlink(file.path(tempdir(), the_test))