Load WORCS project data

Scans the WORCS project file for data that have been saved using open_data or closed_data, and loads these data into the global (working) environment. The function will load the original data if available on the current system. If only a synthetic dataset is available, this function loads the synthetic data. The name of the object containing the data is derived from the file name by removing the file extension, and, when applicable, the prefix "synthetic_". Thus, both "data.csv" and "synthetic_data.csv" will be loaded into an object called data.

load_data(worcs_directory = ".", to_envir = TRUE, envir = parent.frame(1))

Arguments

worcs_directory
Character, indicating the WORCS project directory from which to load data. The default value "." points to the current directory.

to_envir
Logical, indicating whether to load objects directly into the environment, or return a list containing the objects. The environment is designated by argument envir. Loading objects directly into the global environment is user-friendly, but has the risk of overwriting an existing object with the same name, as explained in load. The function load_data gives a warning when this happens.

envir
The environment where the data should be loaded. The default value parent.frame(1) refers to the global environment in an interactive session.

Value

Returns a list invisibly. If to_envir = TRUE, this list contains the loaded data files. If to_envir = FALSE, the list is empty, and the loaded data files are attached directly to the global environment.

Examples

test_dir <- file.path(tempdir(), "loaddata")
old_wd <- getwd()
dir.create(test_dir)
setwd(test_dir)
worcs:::write_worcsfile(".worcs")
suppressWarnings(closed_data(iris[1:5, ], codebook = NULL))

#> v  Storing original data in 'data.csv' and updating the checksum in '.worcs'.
#> v  Generating synthetic data for public use. Ensure that no identifying information is included.
#> v  Storing synthetic data in 'synthetic_data.csv' and updating the checksum in '.worcs'.
#> v  Updating '.gitignore'.

load_data()

#> v  Loading original data from 'data.csv'.

data

#> This is the original data set. The first 6 rows are:
#> 
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa

rm("data")
file.remove("data.csv")

#> [1] TRUE

load_data()

#> v  Loading synthetic data from 'synthetic_data.csv'.

data

#> This is a synthetic data set. The first 6 rows are:
#> 
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.0      3.2788      1.39968         0.2  setosa
#> 2          4.9      3.2788      1.39968         0.2  setosa
#> 3          4.6      3.2788      1.39968         0.2  setosa
#> 4          4.9      3.2788      1.39968         0.2  setosa
#> 5          4.7      3.2788      1.39968         0.2  setosa

setwd(old_wd)
unlink(test_dir, recursive = TRUE)