Create new WORCS project

Creates a new 'worcs' project. This function is invoked by the 'RStudio' project template manager, but can also be called directly to create a WORCS project through syntax or the console.

worcs_project(
  path = "worcs_project",
  manuscript = "APA6",
  preregistration = "COS",
  add_license = "CC_BY_4.0",
  use_renv = TRUE,
  remote_repo = "https",
  ...
)

Arguments

path
Character, indicating the directory in which to create the 'worcs' project. Default: 'worcs_project'.

manuscript
Character, indicating what template to use for the 'rmarkdown' manuscript. Default: 'APA6'. Available choices include: "APA6", "github_document", "None", "ams_article", "asa_article", "biometrics_article", "copernicus_article", "ctex", "elsevier_article", "frontiers_article", "ieee_article", "joss_article", "jss_article", "mdpi_article", "mnras_article", "oup_article", "peerj_article", "plos_article", "pnas_article", "rjournal_article", "rsos_article", "sage_article", "sim_article", "springer_article", "tf_article". For more information about APA6, see the 'papaja' package, at <https://github.com/crsh/papaja>. For more information about github_document, see github_document. For the remaining formats, see, e.g., acm_article.

preregistration
Character, indicating what template to use for the preregistration. Default: 'COS'. Available choices include: "COS", "VantVeer", "Brandt", "AsPredicted", "None". For more information, see, e.g., cos_prereg.

add_license
Character, indicating what license to include. Default: 'CC_BY_4.0'. Available options include: "CC_BY_4.0", "CC_BY-SA_4.0", "CC_BY-NC_4.0", "CC_BY-NC-SA_4.0", "CC_BY-ND_4.0", "CC_BY-NC-ND_4.0", "None". For more information, see <https://creativecommons.org/licenses/>.

use_renv
Logical, indicating whether or not to use 'renv' to make the project reproducible. Default: TRUE. See init.

remote_repo
Character, 'https' link to the 'GitHub' repository for this project. If a valid 'GitHub' repository link is provided, a commit will be made containing the 'README.md' file, and will be pushed to 'GitHub'. Default: 'https'. For more information, see <http://github.com/>.

...
Additional arguments passed to and from functions.

Value

No return value. This function is called for its side effects.

Examples

the_test <- "worcs_template"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
get_sig <- tryCatch(git_signature_default(), error = function(e){
  gert::git_config_global_set(name = "user.name", value = "yourname")
  gert::git_config_global_set(name = "user.email", value = "yourname@email.com")
})
worcs_project(file.path(tempdir(), the_test, "worcs_project"),
              manuscript = "github_document",
              preregistration = "None",
              add_license = "None",
              use_renv = FALSE,
              remote_repo = "https")

#> Called from: worcs_project(file.path(tempdir(), the_test, "worcs_project"), 
#>     manuscript = "github_document", preregistration = "None", 
#>     add_license = "None", use_renv = FALSE, remote_repo = "https")
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#56: manuscript <- tolower(manuscript)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#57: preregistration <- tolower(preregistration)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#58: add_license <- tolower(add_license)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#59: dots <- list(...)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#61: dir.create(path, recursive = TRUE, showWarnings = FALSE)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#62: path <- normalizePath(path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#63: git_init(path = path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#65: write_worcsfile(filename = file.path(path, ".worcs"), worcs_version = as.character(packageVersion("worcs")), 
#>     creator = Sys.info()["effective_user"])
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#71: copy_resources(which_files = c("README.md", "prepare_data.R"), 
#>     path = path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#79: if (!manuscript == "none") {
#>     man_dir <- file.path(path, "manuscript")
#>     manuscript_file <- file.path(man_dir, "manuscript.Rmd")
#>     dir.create(man_dir)
#>     switch(manuscript, apa6 = create_man_papaja(manuscript_file, 
#>         remote_repo = remote_repo), acm_article = create_man_rticles(manuscript_file, 
#>         "acm_article", remote_repo = remote_repo), acs_article = create_man_rticles(manuscript_file, 
#>         "acs_article", remote_repo = remote_repo), aea_article = create_man_rticles(manuscript_file, 
#>         "aea_article", remote_repo = remote_repo), agu_article = create_man_rticles(manuscript_file, 
#>         "agu_article", remote_repo = remote_repo), amq_article = create_man_rticles(manuscript_file, 
#>         "amq_article", remote_repo = remote_repo), ams_article = create_man_rticles(manuscript_file, 
#>         "ams_article", remote_repo = remote_repo), asa_article = create_man_rticles(manuscript_file, 
#>         "asa_article", remote_repo = remote_repo), biometrics_article = create_man_rticles(manuscript_file, 
#>         "biometrics_article", remote_repo = remote_repo), copernicus_article = create_man_rticles(manuscript_file, 
#>         "copernicus_article", remote_repo = remote_repo), ctex = create_man_rticles(manuscript_file, 
#>         "ctex", remote_repo = remote_repo), elsevier_article = create_man_rticles(manuscript_file, 
#>         "elsevier_article", remote_repo = remote_repo), frontiers_article = create_man_rticles(manuscript_file, 
#>         "frontiers_article", remote_repo = remote_repo), ieee_article = create_man_rticles(manuscript_file, 
#>         "ieee_article", remote_repo = remote_repo), joss_article = create_man_rticles(manuscript_file, 
#>         "joss_article", remote_repo = remote_repo), jss_article = create_man_rticles(manuscript_file, 
#>         "jss_article", remote_repo = remote_repo), mdpi_article = create_man_rticles(manuscript_file, 
#>         "mdpi_article", remote_repo = remote_repo), mnras_article = create_man_rticles(manuscript_file, 
#>         "mnras_article", remote_repo = remote_repo), oup_article = create_man_rticles(manuscript_file, 
#>         "oup_article", remote_repo = remote_repo), peerj_article = create_man_rticles(manuscript_file, 
#>         "peerj_article", remote_repo = remote_repo), plos_article = create_man_rticles(manuscript_file, 
#>         "plos_article", remote_repo = remote_repo), pnas_article = create_man_rticles(manuscript_file, 
#>         "pnas_article", remote_repo = remote_repo), rjournal_article = create_man_rticles(manuscript_file, 
#>         "rjournal_article", remote_repo = remote_repo), rsos_article = create_man_rticles(manuscript_file, 
#>         "rsos_article", remote_repo = remote_repo), sage_article = create_man_rticles(manuscript_file, 
#>         "sage_article", remote_repo = remote_repo), sim_article = create_man_rticles(manuscript_file, 
#>         "sim_article", remote_repo = remote_repo), springer_article = create_man_rticles(manuscript_file, 
#>         "springer_article", remote_repo = remote_repo), tf_article = create_man_rticles(manuscript_file, 
#>         "tf_article", remote_repo = remote_repo), create_man_github(manuscript_file, 
#>         remote_repo = remote_repo))
#>     copy_resources(which_files = "references.bib", path = man_dir)
#>     bibfiles <- list.files(path = man_dir, pattern = ".bib$", 
#>         full.names = TRUE)
#>     if (length(bibfiles) > 1) {
#>         worcs_ref <- readLines(bibfiles[endsWith(bibfiles, "references.bib")])
#>         bib_text <- do.call(c, lapply(bibfiles[!endsWith(bibfiles, 
#>             "references.bib")], readLines))
#>         invisible(file.remove(bibfiles))
#>         writeLines(c(worcs_ref, bib_text), file.path(man_dir, 
#>             "references.bib"))
#>     }
#> }
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#81: man_dir <- file.path(path, "manuscript")
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#82: manuscript_file <- file.path(man_dir, "manuscript.Rmd")
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#83: dir.create(man_dir)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#84: switch(manuscript, apa6 = create_man_papaja(manuscript_file, 
#>     remote_repo = remote_repo), acm_article = create_man_rticles(manuscript_file, 
#>     "acm_article", remote_repo = remote_repo), acs_article = create_man_rticles(manuscript_file, 
#>     "acs_article", remote_repo = remote_repo), aea_article = create_man_rticles(manuscript_file, 
#>     "aea_article", remote_repo = remote_repo), agu_article = create_man_rticles(manuscript_file, 
#>     "agu_article", remote_repo = remote_repo), amq_article = create_man_rticles(manuscript_file, 
#>     "amq_article", remote_repo = remote_repo), ams_article = create_man_rticles(manuscript_file, 
#>     "ams_article", remote_repo = remote_repo), asa_article = create_man_rticles(manuscript_file, 
#>     "asa_article", remote_repo = remote_repo), biometrics_article = create_man_rticles(manuscript_file, 
#>     "biometrics_article", remote_repo = remote_repo), copernicus_article = create_man_rticles(manuscript_file, 
#>     "copernicus_article", remote_repo = remote_repo), ctex = create_man_rticles(manuscript_file, 
#>     "ctex", remote_repo = remote_repo), elsevier_article = create_man_rticles(manuscript_file, 
#>     "elsevier_article", remote_repo = remote_repo), frontiers_article = create_man_rticles(manuscript_file, 
#>     "frontiers_article", remote_repo = remote_repo), ieee_article = create_man_rticles(manuscript_file, 
#>     "ieee_article", remote_repo = remote_repo), joss_article = create_man_rticles(manuscript_file, 
#>     "joss_article", remote_repo = remote_repo), jss_article = create_man_rticles(manuscript_file, 
#>     "jss_article", remote_repo = remote_repo), mdpi_article = create_man_rticles(manuscript_file, 
#>     "mdpi_article", remote_repo = remote_repo), mnras_article = create_man_rticles(manuscript_file, 
#>     "mnras_article", remote_repo = remote_repo), oup_article = create_man_rticles(manuscript_file, 
#>     "oup_article", remote_repo = remote_repo), peerj_article = create_man_rticles(manuscript_file, 
#>     "peerj_article", remote_repo = remote_repo), plos_article = create_man_rticles(manuscript_file, 
#>     "plos_article", remote_repo = remote_repo), pnas_article = create_man_rticles(manuscript_file, 
#>     "pnas_article", remote_repo = remote_repo), rjournal_article = create_man_rticles(manuscript_file, 
#>     "rjournal_article", remote_repo = remote_repo), rsos_article = create_man_rticles(manuscript_file, 
#>     "rsos_article", remote_repo = remote_repo), sage_article = create_man_rticles(manuscript_file, 
#>     "sage_article", remote_repo = remote_repo), sim_article = create_man_rticles(manuscript_file, 
#>     "sim_article", remote_repo = remote_repo), springer_article = create_man_rticles(manuscript_file, 
#>     "springer_article", remote_repo = remote_repo), tf_article = create_man_rticles(manuscript_file, 
#>     "tf_article", remote_repo = remote_repo), create_man_github(manuscript_file, 
#>     remote_repo = remote_repo))
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#116: copy_resources(which_files = "references.bib", path = man_dir)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#117: bibfiles <- list.files(path = man_dir, pattern = ".bib$", full.names = TRUE)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#118: if (length(bibfiles) > 1) {
#>     worcs_ref <- readLines(bibfiles[endsWith(bibfiles, "references.bib")])
#>     bib_text <- do.call(c, lapply(bibfiles[!endsWith(bibfiles, 
#>         "references.bib")], readLines))
#>     invisible(file.remove(bibfiles))
#>     writeLines(c(worcs_ref, bib_text), file.path(man_dir, "references.bib"))
#> }
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#129: if (!preregistration == "none") {
#>     draft(file.path(path, "preregistration.Rmd"), paste0(preregistration, 
#>         "_prereg"), package = "prereg", create_dir = FALSE, edit = FALSE)
#> }
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#142: if (!add_license == "none") {
#>     dir.create(path, recursive = TRUE, showWarnings = FALSE)
#>     license_dir = system.file("rstudio", "templates", "project", 
#>         "licenses", package = "worcs", mustWork = TRUE)
#>     license_file <- file.path(license_dir, paste0(add_license, 
#>         ".txt"))
#>     if (file.exists(license_file)) {
#>         file.copy(license_file, file.path(path, "LICENSE"))
#>     }
#>     else {
#>         warning("Could not find requested license.")
#>     }
#> }
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#157: if (use_renv) {
#>     init_fun <- get("init", asNamespace("renv"))
#>     do.call(init_fun, list(project = path, restart = FALSE))
#> }
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#164: write(c(".Rhistory", ".Rprofile", "*.csv", "*.sav", "*.sas7bdat", 
#>     "*.xlsx", "*.xls", "*.pdf", "*.fff", "*.log", "*.tex"), file = file.path(path, 
#>     ".gitignore"), append = TRUE)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#178: if (file.exists(file.path(path, "README.md"))) {
#>     cont <- readLines(file.path(path, "README.md"))
#>     f <- list.files(path)
#>     tab <- matrix(c("File", "Description", "Usage", "README.md", 
#>         "Description of project", "Human editable"), nrow = 2, 
#>         byrow = TRUE)
#>     rproj_name <- paste0(gsub("^.+\\b(.+)$", "\\1", path), ".Rproj")
#>     cont[grep("You can load this project in Rstudio by opening the file called ", 
#>         cont)] <- paste0(grep("You can load this project in Rstudio by opening the file called ", 
#>         cont, value = TRUE), "'", rproj_name, "'.")
#>     tab <- rbind(tab, c(rproj_name, "Project file", "Loads project"))
#>     tab <- describe_file("LICENSE", "User permissions", "Read only", 
#>         tab, path)
#>     tab <- describe_file("manuscript.rmd", "Source code for paper", 
#>         "Human editable", tab, path)
#>     tab <- describe_file("preregistration.rmd", "Preregistered hypotheses", 
#>         "Human editable", tab, path)
#>     tab <- describe_file("prepare_data.R", "Script to process raw data", 
#>         "Human editable", tab, path)
#>     tab <- describe_file("renv.lock", "Reproducible R environment", 
#>         "Read only", tab, path)
#>     tab <- append(apply(tab, 1, paste, collapse = " | "), "--- | --- | ---", 
#>         after = 1)
#>     cont <- append(cont, tab, after = grep("You can add rows to this table", 
#>         cont))
#>     writeLines(cont, file.path(path, "README.md"))
#> }
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#179: cont <- readLines(file.path(path, "README.md"))
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#180: f <- list.files(path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#181: tab <- matrix(c("File", "Description", "Usage", "README.md", 
#>     "Description of project", "Human editable"), nrow = 2, byrow = TRUE)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#183: rproj_name <- paste0(gsub("^.+\\b(.+)$", "\\1", path), ".Rproj")
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#184: cont[grep("You can load this project in Rstudio by opening the file called ", 
#>     cont)] <- paste0(grep("You can load this project in Rstudio by opening the file called ", 
#>     cont, value = TRUE), "'", rproj_name, "'.")
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#185: tab <- rbind(tab, c(rproj_name, "Project file", "Loads project"))
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#186: tab <- describe_file("LICENSE", "User permissions", "Read only", 
#>     tab, path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#187: tab <- describe_file("manuscript.rmd", "Source code for paper", 
#>     "Human editable", tab, path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#188: tab <- describe_file("preregistration.rmd", "Preregistered hypotheses", 
#>     "Human editable", tab, path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#189: tab <- describe_file("prepare_data.R", "Script to process raw data", 
#>     "Human editable", tab, path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#190: tab <- describe_file("renv.lock", "Reproducible R environment", 
#>     "Read only", tab, path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#191: tab <- append(apply(tab, 1, paste, collapse = " | "), "--- | --- | ---", 
#>     after = 1)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#192: cont <- append(cont, tab, after = grep("You can add rows to this table", 
#>     cont))
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#193: writeLines(cont, file.path(path, "README.md"))
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#197: git_add(files = "README.md", repo = path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#198: git_commit(message = "worcs template initial commit", repo = path)
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#201: if (grepl("^https://github.com/.+?/.+?\\.git$", remote_repo)) {
#>     tryCatch({
#>         git_remote_add(name = "origin", url = remote_repo, repo = path)
#>         git_push(remote = "origin", repo = path)
#>     }, error = function(e) {
#>         message("Could not connect to a remote 'GitHub' repository. You are working with a local 'Git' repository only.")
#>     })
#> } else {
#>     message("No valid 'GitHub' address provided. You are working with a local 'Git' repository only.")
#> }
#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#207: message("No valid 'GitHub' address provided. You are working with a local 'Git' repository only.")

#> No valid 'GitHub' address provided. You are working with a local 'Git' repository only.

#> debug at C:/Git_Repositories/worcs/R/worcs_template.R#209: if ("GCtorture" %in% ls()) rm("GCtorture")

setwd(old_wd)
unlink(file.path(tempdir(), the_test))