Create new WORCS project

Creates a new 'worcs' project. This function is invoked by the 'RStudio' project template manager, but can also be called directly to create a WORCS project through syntax or the console.

Usage

worcs_project(
  path = "worcs_project",
  manuscript = "APA6",
  preregistration = "cos_prereg",
  add_license = "CC_BY_4.0",
  use_renv = TRUE,
  use_targets = FALSE,
  remote_repo = "https",
  verbose = TRUE,
  ...
)

Arguments

path

Character, indicating the directory in which to create the 'worcs' project. Default: 'worcs_project'.

manuscript

Character, indicating what template to use for the 'R Markdown' manuscript. Default: 'APA6'. Available choices include APA6 from the papaja package, a github_document, and templates included in the rticles package. For more information, see add_manuscript.

preregistration

Character, indicating what template to use for the preregistration. Default: 'cos_prereg'. Available choices include: "PSS", "Secondary", "None", and all templates from the prereg package. For more information, see add_preregistration.

add_license

Character, indicating what license to include. Default: 'CC_BY_4.0'. Available options include: "CC_BY_4.0", "CC_BY-SA_4.0", "CC_BY-NC_4.0", "CC_BY-NC-SA_4.0", "CC_BY-ND_4.0", "CC_BY-NC-ND_4.0", "None". For more information, see https://creativecommons.org/licenses/.

use_renv

Logical, indicating whether or not to use 'renv' to make the project reproducible. Default: TRUE. See init.

use_targets

Logical, indicating whether or not to use 'targets' to create a Make-like pipeline. Default: FALSE See targets-package.

remote_repo

Character, address of the remote repository for this project. This link should have the form https://github.com[username][repo].git (preferred) or git@[...].git (if using SSH). If a valid remote repository link is provided, a commit will be made containing the 'README.md' file, and will be pushed to the remote repository. Default: 'https'.

verbose

Logical. Whether or not to print messages to the console during project creation. Default: TRUE

...

Additional arguments passed to and from functions.

Value

No return value. This function is called for its side effects.

Examples

the_test <- "worcs_template"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
do.call(git_user, worcs:::get_user())
#> To set the 'Git' username and email, call 'git_user()' with the argument 'overwrite = TRUE'.
worcs_project(file.path(tempdir(), the_test, "worcs_project"),
              manuscript = "github_document",
              preregistration = "None",
              add_license = "None",
              use_renv = FALSE,
              remote_repo = "https")
#> Called from: worcs_project(file.path(tempdir(), the_test, "worcs_project"), 
#>     manuscript = "github_document", preregistration = "None", 
#>     add_license = "None", use_renv = FALSE, remote_repo = "https")
#> debug at C:/git_repositories/worcs/R/worcs_project.R#68: cl <- match.call(expand.dots = FALSE)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#71: manuscript <- tolower(manuscript)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#73: if (manuscript == "target_markdown") use_targets <- TRUE
#> debug at C:/git_repositories/worcs/R/worcs_project.R#74: preregistration <- tolower(preregistration)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#75: add_license <- tolower(add_license)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#76: dots <- list(...)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#78: dir.create(path, recursive = TRUE, showWarnings = FALSE)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#79: path <- normalizePath(path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#81: use_git <- has_git()
#> debug at C:/git_repositories/worcs/R/worcs_project.R#82: if (!use_git) {
#>     col_message("Could not find a working installation of 'Git', which is required to safeguard the transparency and reproducibility of your project. Please connect 'Git' by following the steps described in this vignette:\n  vignette('setup', package = 'worcs')", 
#>         success = FALSE)
#> } else {
#>     col_message("Initializing 'Git' repository.", verbose = verbose)
#>     git_init(path = path)
#> }
#> debug at C:/git_repositories/worcs/R/worcs_project.R#85: col_message("Initializing 'Git' repository.", verbose = verbose)
#> ✔ Initializing 'Git' repository.
#> debug at C:/git_repositories/worcs/R/worcs_project.R#86: git_init(path = path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#90: tryCatch({
#>     write_worcsfile(filename = file.path(path, ".worcs"), worcs_version = as.character(packageVersion("worcs")), 
#>         creator = Sys.info()["effective_user"])
#>     col_message("Writing '.worcs' file.", verbose = verbose)
#> }, error = function(e) {
#>     col_message("Writing '.worcs' file.", success = FALSE)
#> })
#> debug at C:/git_repositories/worcs/R/worcs_project.R#91: write_worcsfile(filename = file.path(path, ".worcs"), worcs_version = as.character(packageVersion("worcs")), 
#>     creator = Sys.info()["effective_user"])
#> debug at C:/git_repositories/worcs/R/worcs_project.R#95: col_message("Writing '.worcs' file.", verbose = verbose)
#> ✔ Writing '.worcs' file.
#> debug at C:/git_repositories/worcs/R/worcs_project.R#102: tryCatch({
#>     copy_resources(which_files = c("README.md", "prepare_data.R", 
#>         "worcs_icon.png"), path = path)
#>     col_message("Copying standard files.", verbose = verbose)
#> }, error = function(e) {
#>     col_message("Copying standard files.", success = FALSE)
#> })
#> debug at C:/git_repositories/worcs/R/worcs_project.R#103: copy_resources(which_files = c("README.md", "prepare_data.R", 
#>     "worcs_icon.png"), path = path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#108: col_message("Copying standard files.", verbose = verbose)
#> ✔ Copying standard files.
#> debug at C:/git_repositories/worcs/R/worcs_project.R#117: switch(manuscript, none = {
#>     write_as_utf(recommend_data, file.path(path, "run_me.R"))
#>     write_worcsfile(filename = file.path(path, ".worcs"), entry_point = "run_me.R", 
#>         modify = TRUE)
#>     add_recipe(worcs_directory = path, recipe = "source('run_me.R')")
#> }, {
#>     cl[[1L]] <- quote(worcs::add_manuscript)
#>     names(cl)[which(names(cl) == "path")] <- "worcs_directory"
#>     eval(cl, parent.frame())
#>     add_recipe(worcs_directory = path)
#> })
#> debug at C:/git_repositories/worcs/R/worcs_project.R#130: cl[[1L]] <- quote(worcs::add_manuscript)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#131: names(cl)[which(names(cl) == "path")] <- "worcs_directory"
#> debug at C:/git_repositories/worcs/R/worcs_project.R#132: eval(cl, parent.frame())
#> ✔ Creating manuscript files.
#> debug at C:/git_repositories/worcs/R/worcs_project.R#133: add_recipe(worcs_directory = path)
#> ✔ Adding recipe to '.worcs'.
#> debug at C:/git_repositories/worcs/R/worcs_project.R#140: if (!preregistration == "none") {
#>     cl[[1L]] <- quote(worcs::add_preregistration)
#>     names(cl)[which(names(cl) == "path")] <- "worcs_directory"
#>     eval(cl, parent.frame())
#> }
#> debug at C:/git_repositories/worcs/R/worcs_project.R#148: if (!add_license == "none") {
#>     tryCatch({
#>         dir.create(path, recursive = TRUE, showWarnings = FALSE)
#>         license_dir = system.file("rstudio", "templates", "project", 
#>             "licenses", package = "worcs", mustWork = TRUE)
#>         license_file <- file.path(license_dir, paste0(add_license, 
#>             ".txt"))
#>         file.copy(license_file, file.path(path, "LICENSE"), copy.mode = FALSE)
#>         col_message("Writing license file.", verbose = verbose)
#>     }, error = function(e) {
#>         col_message("Writing license file.", success = FALSE)
#>     })
#> }
#> debug at C:/git_repositories/worcs/R/worcs_project.R#165: if (use_targets) {
#>     if (require("targets", quietly = TRUE) & require("tarchetypes", 
#>         quietly = TRUE)) {
#>         worcs::add_targets(worcs_directory = path, verbose = verbose)
#>         col_message("Initializing 'targets' for a Make-like pipeline.", 
#>             verbose = verbose)
#>     }
#> }
#> debug at C:/git_repositories/worcs/R/worcs_project.R#179: if (use_renv) {
#>     tryCatch({
#>         init_fun <- get("init", asNamespace("renv"))
#>         do.call(init_fun, list(project = path, restart = FALSE))
#>         col_message("Initializing 'renv' for a reproducible R environment.", 
#>             verbose = verbose)
#>     }, error = function(e) {
#>         col_message("Initializing 'renv' for a reproducible R environment.", 
#>             success = FALSE)
#>     })
#> }
#> debug at C:/git_repositories/worcs/R/worcs_project.R#191: write(c(".Rhistory", ".Rprofile", "*.csv", "*.sav", "*.sas7bdat", 
#>     "*.xlsx", "*.xls", "*.pdf", "*.fff", "*.log", "*.tex"), file = file.path(path, 
#>     ".gitignore"), append = TRUE)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#205: if (file.exists(file.path(path, "README.md"))) {
#>     cont <- readLines(file.path(path, "README.md"), encoding = "UTF-8")
#>     f <- list.files(path)
#>     tab <- matrix(c("File", "Description", "Usage", "README.md", 
#>         "Description of project", "Human editable"), nrow = 2, 
#>         byrow = TRUE)
#>     rproj_name <- paste0(basename(path), ".Rproj")
#>     cont[which(startsWith(cont, "You can load this project in RStudio by opening the file"))] <- paste0("You can load this project in RStudio by opening the file called '", 
#>         rproj_name, "'.")
#>     tab <- rbind(tab, c(rproj_name, "Project file", "Loads project"))
#>     tab <- describe_file("LICENSE", "User permissions", "Read only", 
#>         tab, path)
#>     tab <- describe_file(".worcs", "WORCS metadata YAML", "Read only", 
#>         tab, path)
#>     tab <- describe_file("preregistration.rmd", "Preregistered hypotheses", 
#>         "Human editable", tab, path)
#>     tab <- describe_file("prepare_data.R", "Script to process raw data", 
#>         "Human editable", tab, path)
#>     tab <- describe_file("manuscript/manuscript.rmd", "Source code for paper", 
#>         "Human editable", tab, path)
#>     tab <- describe_file("manuscript/references.bib", "BibTex references for manuscript", 
#>         "Human editable", tab, path)
#>     tab <- describe_file("renv.lock", "Reproducible R environment", 
#>         "Read only", tab, path)
#>     tab <- nice_tab(tab)
#>     cont <- append(cont, tab, after = grep("You can add rows to this table", 
#>         cont))
#>     write_as_utf(cont, file.path(path, "README.md"))
#> }
#> debug at C:/git_repositories/worcs/R/worcs_project.R#206: cont <- readLines(file.path(path, "README.md"), encoding = "UTF-8")
#> debug at C:/git_repositories/worcs/R/worcs_project.R#207: f <- list.files(path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#208: tab <- matrix(c("File", "Description", "Usage", "README.md", 
#>     "Description of project", "Human editable"), nrow = 2, byrow = TRUE)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#210: rproj_name <- paste0(basename(path), ".Rproj")
#> debug at C:/git_repositories/worcs/R/worcs_project.R#211: cont[which(startsWith(cont, "You can load this project in RStudio by opening the file"))] <- paste0("You can load this project in RStudio by opening the file called '", 
#>     rproj_name, "'.")
#> debug at C:/git_repositories/worcs/R/worcs_project.R#212: tab <- rbind(tab, c(rproj_name, "Project file", "Loads project"))
#> debug at C:/git_repositories/worcs/R/worcs_project.R#213: tab <- describe_file("LICENSE", "User permissions", "Read only", 
#>     tab, path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#214: tab <- describe_file(".worcs", "WORCS metadata YAML", "Read only", 
#>     tab, path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#215: tab <- describe_file("preregistration.rmd", "Preregistered hypotheses", 
#>     "Human editable", tab, path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#216: tab <- describe_file("prepare_data.R", "Script to process raw data", 
#>     "Human editable", tab, path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#217: tab <- describe_file("manuscript/manuscript.rmd", "Source code for paper", 
#>     "Human editable", tab, path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#218: tab <- describe_file("manuscript/references.bib", "BibTex references for manuscript", 
#>     "Human editable", tab, path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#219: tab <- describe_file("renv.lock", "Reproducible R environment", 
#>     "Read only", tab, path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#221: tab <- nice_tab(tab)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#222: cont <- append(cont, tab, after = grep("You can add rows to this table", 
#>     cont))
#> debug at C:/git_repositories/worcs/R/worcs_project.R#223: write_as_utf(cont, file.path(path, "README.md"))
#> debug at C:/git_repositories/worcs/R/worcs_project.R#227: if (use_git) {
#>     tryCatch({
#>         git_add(files = "README.md", repo = path)
#>         git_commit(message = "worcs template initial commit", 
#>             repo = path)
#>         col_message("Creating first commit (committing README.md).", 
#>             verbose = verbose)
#>     }, error = function(e) {
#>         col_message("Creating first commit (committing README.md).", 
#>             success = FALSE)
#>     })
#> }
#> debug at C:/git_repositories/worcs/R/worcs_project.R#228: tryCatch({
#>     git_add(files = "README.md", repo = path)
#>     git_commit(message = "worcs template initial commit", repo = path)
#>     col_message("Creating first commit (committing README.md).", 
#>         verbose = verbose)
#> }, error = function(e) {
#>     col_message("Creating first commit (committing README.md).", 
#>         success = FALSE)
#> })
#> debug at C:/git_repositories/worcs/R/worcs_project.R#229: git_add(files = "README.md", repo = path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#230: git_commit(message = "worcs template initial commit", repo = path)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#231: col_message("Creating first commit (committing README.md).", 
#>     verbose = verbose)
#> ✔ Creating first commit (committing README.md).
#> debug at C:/git_repositories/worcs/R/worcs_project.R#238: repo_url <- parse_repo(remote_repo = remote_repo, verbose = verbose)
#> ✖ Not a valid 'Git' remote repository address: https
#> debug at C:/git_repositories/worcs/R/worcs_project.R#239: valid_repo <- !is.null(repo_url)
#> debug at C:/git_repositories/worcs/R/worcs_project.R#240: if (use_git & valid_repo) {
#>     tryCatch({
#>         Args_gert <- list("origin", url = remote_repo, repo = path)
#>         if ("remote" %in% formalArgs(git_remote_add)) {
#>             names(Args_gert)[1] <- "remote"
#>         }
#>         else {
#>             names(Args_gert)[1] <- "name"
#>         }
#>         do.call(git_remote_add, Args_gert)
#>         git_push(remote = "origin", repo = path)
#>         col_message(paste0("Connected to remote repository at ", 
#>             remote_repo), verbose = verbose)
#>     }, error = function(e) {
#>         col_message("Could not connect to a remote 'GitHub' repository. You are working with a local 'Git' repository only.", 
#>             success = FALSE, verbose = verbose)
#>     })
#> } else {
#>     col_message("No valid 'GitHub' address provided. You are working with a local 'Git' repository only.", 
#>         success = FALSE)
#> }
#> debug at C:/git_repositories/worcs/R/worcs_project.R#260: col_message("No valid 'GitHub' address provided. You are working with a local 'Git' repository only.", 
#>     success = FALSE)
#> ✖ No valid 'GitHub' address provided. You are working with a local 'Git' repository only.
#> debug at C:/git_repositories/worcs/R/worcs_project.R#262: if ("GCtorture" %in% ls()) rm("GCtorture")
setwd(old_wd)
unlink(file.path(tempdir(), the_test))